Separated guess recording and resetting
