+++ lib/delay.ts
@@ -1,0 +1,3 @@
+export function delay(delay: number) {
+    return new Promise(res => setTimeout(res, delay))
+}
